fix keyboard input
read cursor pos directly from window

fix keyboard input

fix: delete and numpad enter.

A better solution would be to use the logical key here and do the
mapping to scancode ourself: this would handle the Numpad arrows
correctly.  We can't just check for NumLock, as the KeyCode for this
is used for 'clear' (!) on Mac.

On the other hand, do emacs users ever use the numpad for anything
other than typing numbers?  I know it highly annoys me when numlock is
off.

fix: mouse position when scale != 1.

input fix