fix: delete and numpad enter.
A better solution would be to use the logical key here and do the
mapping to scancode ourself: this would handle the Numpad arrows
correctly.  We can't just check for NumLock, as the KeyCode for this
is used for 'clear' (!) on Mac.

On the other hand, do emacs users ever use the numpad for anything
other than typing numbers?  I know it highly annoys me when numlock is
off.